bug fixed at adding and editing

+++ mylibrary/src/main/java/com/alina/mylibrary/service/impl/Admin/CategoryServiceImp.java
@@ -1,6 +1,7 @@
 package com.alina.mylibrary.service.impl.Admin;
 
 import com.alina.mylibrary.dao.Interfaces.Admin.CategoryDao;
+import com.alina.mylibrary.exception.ServiceExceptions.DBExceptions;
 import com.alina.mylibrary.model.Category;
 import com.alina.mylibrary.service.Interfaces.Admin.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,12 +18,12 @@ public class CategoryServiceImp implements CategoryService {
 
 
     @Override
-    public Category addCategory(Category category) {
+    public Category addCategory(Category category) throws DBExceptions{
         if (category != null) {
 
             List< com.alina.mylibrary.model.Category> categoriesWithSameTitle=this.categoryDao.getCategoriesWithcategoryTitle(category.getCategoryTitle());
             if(categoriesWithSameTitle.size()>=1){
-                return null;
+               throw new DBExceptions("Categoria este deja inregistrata",1,category.getClass().getName(),"categoryTitle","Insert");
                 //the you already have that category
             }
             return categoryDao.addCategory(category);
@@ -36,12 +37,30 @@ public class CategoryServiceImp implements CategoryService {
     }
 
     @Override
-    public Boolean deleteCategory(int categoryId) {
-       return this.categoryDao.deleteCategory(categoryId);
+    public Boolean deleteCategory(int categoryId) throws DBExceptions {
+        if(categoryId<1){
+            throw new DBExceptions("id-ul nu poate fi gasit",2,"categoryClass","categoryId","Delete");
+        }
+        if(this.categoryDao.findById(categoryId)==false){
+            throw new DBExceptions("id-ul nu poate fi gasit",2,"categoryClass","categoryId","Delete");
+        }
+        else {
+            return this.categoryDao.deleteCategory(categoryId);
+        }
     }
 
     @Override
-    public Category updateCategory(Category category) {
-        return this.categoryDao.updateCategory(category);
+    public Category updateCategory(Category category) throws DBExceptions {
+        if (category != null) {
+
+            List< com.alina.mylibrary.model.Category> categoriesWithSameTitle=this.categoryDao.getCategoriesWithcategoryTitle(category.getCategoryTitle());
+            if(categoriesWithSameTitle.size()>=1){
+                throw new DBExceptions("Categoria este deja inregistrata",1,category.getClass().getName(),"categoryTitle","Insert");
+                //the you already have that category
+            }
+            return categoryDao.updateCategory(category);
+        }
+        return null;
     }
-}
+    }
+
